Membership management, offline access and more
I added membership tracking so that when first app start up, it will prompt the user whether is a member or not.

If the user is a member, it will show a some settings for membership tracking; such as sign-up date and membership type. Also, there will be notifications alerting the user to register when it is close to the deadline. There is now a membership card showing the details of the membership.

If the person is not a member, there will not be a membership card but there will always be the option to register

Offline access is now fixed too, telling the user if he/she is not connected to the internet and showing the cached contents

Other quality of life changes