Reverted FeedView changes
I reverted the changes made to FeedView, this is because I feel that it is still too buggy. I plan to improve my swiftui/swift knowledge before attempting to do this. In the meantime, I will focus on other parts of the app such as ideating the possibility of a homescreen.